+++ internal/schemautil/service.go
@@ -306,10 +306,6 @@ func ResourceServiceCreateWrapper(serviceType string) schema.CreateContextFunc {
 		if err := d.Set("service_type", serviceType); err != nil {
 			return diag.Errorf("error setting service_type: %s", err)
 		}
-		if err := d.Set(serviceType, []map[string]interface{}{}); err != nil {
-			return diag.Errorf("error setting an empty %s field: %s", serviceType, err)
-		}
-
 		return resourceServiceCreate(ctx, d, m)
 	}
 }
@@ -746,17 +742,17 @@ func copyConnectionInfoFromAPIResponseToTerraform(
 	props := make(map[string]interface{})
 
 	switch serviceType {
-	case "opensearch":
+	case ServiceTypeOpenSearch:
 		props["opensearch_dashboards_uri"] = connectionInfo.OpensearchDashboardsURI
-	case "influxdb":
+	case ServiceTypeInfluxDB:
 		props["database_name"] = connectionInfo.InfluxDBDatabaseName
-	case "kafka":
+	case ServiceTypeKafka:
 		props["access_cert"] = connectionInfo.KafkaAccessCert
 		props["access_key"] = connectionInfo.KafkaAccessKey
 		props["connect_uri"] = connectionInfo.KafkaConnectURI
 		props["rest_uri"] = connectionInfo.KafkaRestURI
 		props["schema_registry_uri"] = connectionInfo.SchemaRegistryURI
-	case "pg":
+	case ServiceTypePG:
 		if connectionInfo.PostgresURIs != nil && len(connectionInfo.PostgresURIs) > 0 {
 			props["uri"] = connectionInfo.PostgresURIs[0]
 		}
@@ -774,8 +770,11 @@ func copyConnectionInfoFromAPIResponseToTerraform(
 		}
 		props["replica_uri"] = connectionInfo.PostgresReplicaURI
 		props["max_connections"] = metadata.(map[string]interface{})["max_connections"]
-	case "flink":
+	case ServiceTypeFlink:
 		props["host_ports"] = connectionInfo.FlinkHostPorts
+	default:
+		// Doesn't have connection info
+		return nil
 	}
 
 	return d.Set(serviceType, []map[string]interface{}{props})