+++ internal/sdkprovider/service/grafana/grafana.go
@@ -9,16 +9,7 @@ import (
 )
 
 func grafanaSchema() map[string]*schema.Schema {
-	s := schemautil.ServiceCommonSchemaWithUserConfig(schemautil.ServiceTypeGrafana)
-	s[schemautil.ServiceTypeGrafana] = &schema.Schema{
-		Type:        schema.TypeList,
-		Computed:    true,
-		Description: "Grafana server provided values",
-		Elem: &schema.Resource{
-			Schema: map[string]*schema.Schema{},
-		},
-	}
-	return s
+	return schemautil.ServiceCommonSchemaWithUserConfig(schemautil.ServiceTypeGrafana)
 }
 
 func ResourceGrafana() *schema.Resource {